First version of MapaPobres.html
The first release or partial accomplishment of the final objective. This maps interactively poverty among all Mexicans and the older population. Still a lot has to be done, in particular having to do with refining the menu items. Also, I have found trouble with the matter of fact that the relevant data is contained within tile layers in the maps, and not on feature layers. There were some nice pop-ups in the original maps that are not showing up here... :P.